arranged classes into models and responses
