--- src/ZendeskApi.Client/Models/SingleOrganizationMembership.cs
@@ -1,10 +1,0 @@
-﻿using Newtonsoft.Json;
-
-namespace ZendeskApi.Client.Models
-{
-    public class SingleOrganizationMembership
-    {
-        [JsonProperty("organization_membership")]
-        public OrganizationMembership OrganizationMembership { get; set; }
-    }
-}
