--- src/ZendeskApi.Client/Models/SingleRequest.cs
@@ -1,10 +1,0 @@
-using Newtonsoft.Json;
-
-namespace ZendeskApi.Client.Models
-{
-    public class SingleRequest
-    {
-        [JsonProperty("request")]
-        public Request Request { get; set; }
-    }
-}
