--- src/ZendeskApi.Client/Models/SingleSatisfactionRating.cs
@@ -1,10 +1,0 @@
-using Newtonsoft.Json;
-
-namespace ZendeskApi.Client.Models
-{
-    public class SingleSatisfactionRating
-    {
-        [JsonProperty("satisfaction_rating")]
-        public SatisfactionRating SatisfactionRating { get; set; }
-    }
-}
