--- src/ZendeskApi.Client/Models/SingleTicketField.cs
@@ -1,10 +1,0 @@
-using Newtonsoft.Json;
-
-namespace ZendeskApi.Client.Models
-{
-    public class SingleTicketField
-    {
-        [JsonProperty("ticket_field")]
-        public TicketField TicketField { get; set; }
-    }
-}
