--- src/ZendeskApi.Client/Models/SingleTicketForm.cs
@@ -1,10 +1,0 @@
-using Newtonsoft.Json;
-
-namespace ZendeskApi.Client.Models
-{
-    public class SingleTicketForm
-    {
-        [JsonProperty("ticket_form")]
-        public TicketForm TicketForm { get; set; }
-    }
-}
