+++ src/ZendeskApi.Client/Models/UserFieldResponse.cs
@@ -2,7 +2,7 @@ using Newtonsoft.Json;
 
 namespace ZendeskApi.Client.Models
 {
-    public class SingleUserField
+    public class UserFieldResponse
     {
         [JsonProperty("user_field")]
         public UserField UserField { get; set; }
