+++ src/ZendeskApi.Client/Requests/TicketsRequest.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using Newtonsoft.Json;
+using ZendeskApi.Client.Models;
 using ZendeskApi.Client.Responses;
 
 namespace ZendeskApi.Client.Requests
@@ -7,6 +8,6 @@ namespace ZendeskApi.Client.Requests
     public class TicketsRequest
     {
         [JsonProperty("tickets")]
-        public IEnumerable<TicketResponse> Item { get; set; }
+        public IEnumerable<Ticket> Item { get; set; }
     }
 }
