+++ src/ZendeskApi.Client/Resources/TicketsResource.cs
@@ -36,7 +36,7 @@ namespace ZendeskApi.Client.Resources
 
 
         #region List Tickets
-        public async Task<IPagination<TicketResponse>> ListAsync(PagerParameters pager = null)
+        public async Task<IPagination<Ticket>> ListAsync(PagerParameters pager = null)
         {
             using (_loggerScope(_logger, "ListAsync"))
             using (var client = _apiClient.CreateClient())
@@ -55,7 +55,7 @@ namespace ZendeskApi.Client.Resources
             }
         }
         
-        public async Task<IPagination<TicketResponse>> ListForOrganizationAsync(long organizationId, PagerParameters pager = null)
+        public async Task<IPagination<Ticket>> ListForOrganizationAsync(long organizationId, PagerParameters pager = null)
         {
             using (_loggerScope(_logger, $"ListForOrganizationAsync({organizationId})"))
             using (var client = _apiClient.CreateClient())
@@ -80,7 +80,7 @@ namespace ZendeskApi.Client.Resources
             }
         }
 
-        public async Task<IPagination<TicketResponse>> ListRequestedByAsync(long userId, PagerParameters pager = null)
+        public async Task<IPagination<Ticket>> ListRequestedByAsync(long userId, PagerParameters pager = null)
         {
             using (_loggerScope(_logger, $"ListRequestedByAsync({userId})"))
             using (var client = _apiClient.CreateClient(string.Format(UserResourceUriFormat, userId)))
@@ -105,7 +105,7 @@ namespace ZendeskApi.Client.Resources
             }
         }
 
-        public async Task<IPagination<TicketResponse>> ListCcdAsync(long userId, PagerParameters pager = null)
+        public async Task<IPagination<Ticket>> ListCcdAsync(long userId, PagerParameters pager = null)
         {
             using (_loggerScope(_logger, $"ListCcdAsync({userId})"))
             using (var client = _apiClient.CreateClient(string.Format(UserResourceUriFormat, userId)))
@@ -130,7 +130,7 @@ namespace ZendeskApi.Client.Resources
             }
         }
 
-        public async Task<IPagination<TicketResponse>> ListAssignedToAsync(long userId, PagerParameters pager = null)
+        public async Task<IPagination<Ticket>> ListAssignedToAsync(long userId, PagerParameters pager = null)
         {
             using (_loggerScope(_logger, $"ListAssignedToAsync({userId})"))
             using (var client = _apiClient.CreateClient(string.Format(UserResourceUriFormat, userId)))
@@ -158,7 +158,7 @@ namespace ZendeskApi.Client.Resources
 
 
         #region Show Tickets
-        public async Task<TicketResponseContainer> GetAsync(long ticketId)
+        public async Task<TicketResponse> GetAsync(long ticketId)
         {
             using (_loggerScope(_logger, $"GetAsync({ticketId})"))
             using (var client = _apiClient.CreateClient(ResourceUri))
@@ -173,11 +173,11 @@ namespace ZendeskApi.Client.Resources
 
                 response.EnsureSuccessStatusCode();
 
-                return await response.Content.ReadAsAsync<TicketResponseContainer>();
+                return await response.Content.ReadAsAsync<TicketResponse>();
             }
         }
 
-        public async Task<IPagination<TicketResponse>> GetAsync(long[] ticketIds, PagerParameters pager = null)
+        public async Task<IPagination<Ticket>> GetAsync(long[] ticketIds, PagerParameters pager = null)
         {
             using (_loggerScope(_logger, $"GetAllAsync({ZendeskFormatter.ToCsv(ticketIds)})"))
             using (var client = _apiClient.CreateClient(ResourceUri))
@@ -193,7 +193,7 @@ namespace ZendeskApi.Client.Resources
 
 
         #region Create Tickets
-        public async Task<TicketResponseContainer> CreateAsync(TicketCreateRequest ticket)
+        public async Task<TicketResponse> CreateAsync(TicketCreateRequest ticket)
         {
             using (_loggerScope(_logger, "CreateAsync"))
             using (var client = _apiClient.CreateClient())
@@ -209,7 +209,7 @@ namespace ZendeskApi.Client.Resources
                                     .Build(); 
                 }
 
-                return await response.Content.ReadAsAsync<TicketResponseContainer>();
+                return await response.Content.ReadAsAsync<TicketResponse>();
             }
         }
 
@@ -236,7 +236,7 @@ namespace ZendeskApi.Client.Resources
 
 
         #region Update Tickets
-        public async Task<TicketResponseContainer> UpdateAsync(TicketUpdateRequest ticket)
+        public async Task<TicketResponse> UpdateAsync(TicketUpdateRequest ticket)
         {
             using (_loggerScope(_logger, "UpdateAsync"))
             using (var client = _apiClient.CreateClient(ResourceUri))
@@ -257,7 +257,7 @@ namespace ZendeskApi.Client.Resources
                                     .Build();
                 }
 
-                return await response.Content.ReadAsAsync<TicketResponseContainer>();
+                return await response.Content.ReadAsAsync<TicketResponse>();
             }
         }
