+++ src/ZendeskApi.Client/Responses/GroupsResponse.cs
@@ -6,11 +6,11 @@ using ZendeskApi.Client.Resources;
 namespace ZendeskApi.Client.Responses
 {
     [JsonObject]
-    public class GroupListResponse : PaginationResponse<GroupResponse>
+    public class GroupListResponse : PaginationResponse<Group>
     {
         [JsonProperty("groups")]
-        public IEnumerable<GroupResponse> Groups { get; set; }
+        public IEnumerable<Group> Groups { get; set; }
         
-        protected override IEnumerable<GroupResponse> Enumerable => Groups;
+        protected override IEnumerable<Group> Enumerable => Groups;
     }
 }
