+++ src/ZendeskApi.Client/Responses/OrganizationMembershipResponse.cs
@@ -1,15 +1,11 @@
-using System.Collections.Generic;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 using ZendeskApi.Client.Models;
 
 namespace ZendeskApi.Client.Responses
 {
-    [JsonObject]
-    public class OrganizationMembershipsResponse : PaginationResponse<OrganizationMembership>
+    public class OrganizationMembershipResponse
     {
-        [JsonProperty("organization_memberships")]
-        public IEnumerable<OrganizationMembership> OrganizationMemberships { get; internal set; }
-        
-        protected override IEnumerable<OrganizationMembership> Enumerable => OrganizationMemberships;
+        [JsonProperty("organization_membership")]
+        public OrganizationMembership OrganizationMembership { get; set; }
     }
 }
