+++ src/ZendeskApi.Client/Responses/SatisfactionRatingResponse.cs
@@ -1,15 +1,11 @@
-using System.Collections.Generic;
 using Newtonsoft.Json;
 using ZendeskApi.Client.Models;
 
 namespace ZendeskApi.Client.Responses
 {
-    [JsonObject]
-    public class SatisfactionRatingsResponse : PaginationResponse<SatisfactionRating>
+    public class SatisfactionRatingResponse
     {
-        [JsonProperty("satisfaction_ratings")]
-        public IEnumerable<SatisfactionRating> SatisfactionRatings { get; set; }
-
-        protected override IEnumerable<SatisfactionRating> Enumerable => SatisfactionRatings;
+        [JsonProperty("satisfaction_rating")]
+        public SatisfactionRating SatisfactionRating { get; set; }
     }
 }
