--- src/ZendeskApi.Client/Responses/SingleGroupResponse.cs
@@ -1,10 +1,0 @@
-using Newtonsoft.Json;
-
-namespace ZendeskApi.Client.Responses
-{
-    public class SingleGroupResponse
-    {
-        [JsonProperty]
-        public GroupResponse Group { get; set; }
-    }
-}
