+++ src/ZendeskApi.Client/Responses/TicketFieldResponse.cs
@@ -1,15 +1,11 @@
-using System.Collections.Generic;
 using Newtonsoft.Json;
 using ZendeskApi.Client.Models;
 
 namespace ZendeskApi.Client.Responses
 {
-    [JsonObject]
-    public class TicketFieldsResponse : PaginationResponse<TicketField>
+    public class TicketFieldResponse
     {
-        [JsonProperty("ticket_fields")]
-        public IEnumerable<TicketField> TicketFields { get; internal set; }
-
-        protected override IEnumerable<TicketField> Enumerable => TicketFields;
+        [JsonProperty("ticket_field")]
+        public TicketField TicketField { get; set; }
     }
 }
