--- src/ZendeskApi.Client/Responses/TicketResponseContainer.cs
@@ -1,10 +1,0 @@
-using Newtonsoft.Json;
-
-namespace ZendeskApi.Client.Responses
-{
-    public class TicketResponseContainer
-    {
-        [JsonProperty("ticket")]
-        public TicketResponse Ticket { get; set; }
-    }
-}
