+++ src/ZendeskApi.Client/Responses/TicketsListResponse.cs
@@ -1,15 +1,16 @@
 using System.Collections.Generic;
 using Newtonsoft.Json;
+using ZendeskApi.Client.Models;
 
 namespace ZendeskApi.Client.Responses
 {
     [JsonObject]
-    public class TicketsListResponse : PaginationResponse<TicketResponse>
+    public class TicketsListResponse : PaginationResponse<Ticket>
     {
         [JsonProperty("tickets")]
-        public IEnumerable<TicketResponse> Tickets { get; set; }
+        public IEnumerable<Ticket> Tickets { get; set; }
 
 
-        protected override IEnumerable<TicketResponse> Enumerable => Tickets;
+        protected override IEnumerable<Ticket> Enumerable => Tickets;
     }
 }
