+++ test/ZendeskApi.Client.Tests/ResourcesSampleSites/TicketFormsResourceSampleSite.cs
@@ -54,7 +54,7 @@ namespace ZendeskApi.Client.Tests.ResourcesSampleSites
                         var obj = state.TicketForms.Single(x => x.Key == id).Value;
 
                         resp.StatusCode = (int)HttpStatusCode.OK;
-                        return resp.WriteAsJson(new SingleTicketForm { TicketForm = obj });
+                        return resp.WriteAsJson(new TicketFormResponse { TicketForm = obj });
                     })
                     .MapGet("api/v2/ticket_forms", (req, resp, routeData) =>
                     {
