+++ test/ZendeskApi.Client.Tests/ResourcesSampleSites/TicketResourceSampleSite.cs
@@ -21,7 +21,7 @@ namespace ZendeskApi.Client.Tests.ResourcesSampleSites
     public class TicketResourceSampleSite : SampleSite
     {
         private class State {
-            public readonly IDictionary<long, TicketResponse> Tickets = new Dictionary<long, TicketResponse>();
+            public readonly IDictionary<long, Ticket> Tickets = new Dictionary<long, Ticket>();
             public readonly IDictionary<long, IList<TicketComment>> TicketComments = new Dictionary<long, IList<TicketComment>>();
         }
          
@@ -62,7 +62,7 @@ namespace ZendeskApi.Client.Tests.ResourcesSampleSites
                         var ticket = state.Tickets.Single(x => x.Key == id).Value;
 
                         resp.StatusCode = (int)HttpStatusCode.OK;
-                        return resp.WriteAsJson(new TicketResponseContainer{Ticket = ticket});
+                        return resp.WriteAsJson(new TicketResponse{Ticket = ticket});
                     })
                     .MapGet("api/v2/tickets", (req, resp, routeData) =>
                     {
@@ -157,7 +157,7 @@ namespace ZendeskApi.Client.Tests.ResourcesSampleSites
                         }
 
                         var mapper = req.HttpContext.RequestServices.GetRequiredService<IMapper>();
-                        var ticketResponse = mapper.Map<TicketResponseContainer>(ticket);
+                        var ticketResponse = mapper.Map<TicketResponse>(ticket);
 
                         var state = req.HttpContext.RequestServices.GetRequiredService<State>();
                         ticketResponse.Ticket.Id = long.Parse(Rand.Next().ToString());
@@ -188,7 +188,7 @@ namespace ZendeskApi.Client.Tests.ResourcesSampleSites
 
                         HandleTicketComment(ticketRequest.Comment, state, ticketRequest.Id);
 
-                        var ticketResponse = new TicketResponseContainer {Ticket = state.Tickets[id] };
+                        var ticketResponse = new TicketResponse {Ticket = state.Tickets[id] };
                         var mapper = req.HttpContext.RequestServices.GetRequiredService<IMapper>();
                         mapper.Map(ticketRequest, ticketResponse);
 
@@ -242,12 +242,12 @@ namespace ZendeskApi.Client.Tests.ResourcesSampleSites
                     services.AddSingleton(_ => new State());
                     services.AddSingleton(_ => new MapperConfiguration(cfg =>
                         {
-                            cfg.CreateMap<TicketCreateRequest, TicketResponseContainer>()
+                            cfg.CreateMap<TicketCreateRequest, TicketResponse>()
                                 .ForMember(r => r.Ticket, r => r.MapFrom(req => req));
-                            cfg.CreateMap<TicketUpdateRequest, TicketResponseContainer>()
+                            cfg.CreateMap<TicketUpdateRequest, TicketResponse>()
                                 .ForMember(r => r.Ticket, r => r.MapFrom(req => req));
-                            cfg.CreateMap<TicketCreateRequest, TicketResponse>();
-                            cfg.CreateMap<TicketUpdateRequest, TicketResponse>();
+                            cfg.CreateMap<TicketCreateRequest, Ticket>();
+                            cfg.CreateMap<TicketUpdateRequest, Ticket>();
                         }).CreateMapper());
                     services.AddRouting();
                     services.AddMemoryCache();
